+++ store/modules/stimulation/actions.js
@@ -1,14 +1,7 @@
 import { WellTitle as LabwareDefinition } from "@/js_utils/labware_calculations.js";
 const twenty_four_well_plate_definition = new LabwareDefinition(4, 6);
 import { call_axios_post_from_vuex } from "../../../js_utils/axios_helpers";
-import { STIM_STATUS } from "./enums";
-
-const TIME_CONVERSION = {
-  milliseconds: 1,
-  seconds: 1000,
-  minutes: 60000,
-  hours: 3600000,
-};
+import { STIM_STATUS, TIME_CONVERSION_TO_MILLIS } from "./enums";
 
 export default {
   handle_selected_wells({ commit }, wells) {
@@ -21,7 +14,7 @@ export default {
     commit("set_selected_wells", well_values);
   },
 
-  async handle_protocol_order({ commit, dispatch }, new_pulse_order) {
+  async handle_protocol_order({ commit, dispatch, state }, new_pulse_order) {
     const x_values = [0];
     const y_values = [0];
     const color_assignments = {};
@@ -66,7 +59,7 @@ export default {
 
       const starting_repeat_idx = x_values.length - 1;
       const converted_total_active =
-        total_active_duration.duration * TIME_CONVERSION[total_active_duration.unit];
+        total_active_duration.duration * TIME_CONVERSION_TO_MILLIS[total_active_duration.unit];
 
       setting = {
         ...setting,
@@ -85,6 +78,11 @@ export default {
       color_assignments[color] = [starting_repeat_idx, ending_repeat_idx];
     });
 
+    // convert x_values to correct unit
+    x_values.forEach((val, idx) => {
+      x_values[idx] = val / TIME_CONVERSION_TO_MILLIS[state.x_axis_unit_name];
+    });
+
     commit("set_repeat_color_assignments", color_assignments);
     commit("set_pulses", { pulses, new_pulse_order });
     dispatch("handle_rest_duration", {
@@ -98,7 +96,7 @@ export default {
     let delay_block;
 
     if (rest_duration !== 0) {
-      const converted_delay = rest_duration * TIME_CONVERSION[time_unit];
+      const converted_delay = rest_duration * TIME_CONVERSION_TO_MILLIS[time_unit];
       const last_x_value = x_values[x_values.length - 1];
       const next_x_value = last_x_value + converted_delay;
       delay_block = [last_x_value, next_x_value];
@@ -333,7 +331,8 @@ export default {
     }
     commit("reset_protocol_editor");
   },
-  handle_x_axis_unit({ commit, dispatch, state }, idx) {
+  handle_x_axis_unit({ commit, dispatch, state }, { idx, unit_name }) {
+    state.x_axis_unit_name = unit_name;
     const { x_axis_values, y_axis_values, x_axis_time_idx } = state;
     if (idx !== x_axis_time_idx) {
       const converted_x_values = x_axis_values.map((val) => (idx === 1 ? val * 1e-3 : val * 1e3));
@@ -350,8 +349,11 @@ export default {
     const well_indices = Object.keys(state.protocol_assignments);
     const res = await call_axios_post_from_vuex(url, { well_indices });
 
-    if (res && res.status !== 200) commit("set_stim_status", STIM_STATUS.ERROR);
-    else commit("set_stim_status", STIM_STATUS.CONFIG_CHECK_IN_PROGRESS);
+    if (res && res.status !== 200) {
+      commit("set_stim_status", STIM_STATUS.ERROR);
+    } else {
+      commit("set_stim_status", STIM_STATUS.CONFIG_CHECK_IN_PROGRESS);
+    }
   },
   async on_pulse_mouseenter({ state }, idx) {
     const hovered_pulse = Object.entries(state.repeat_colors)[idx];